refactor: split main into more modules
